Refactor: begin moving things out of mailpile.search into mailpile.index.*
Note: This commit leaves Mailpile in a non-working state, following
      commits are required to make sure things work.
